extract rsm to common
